refactor(agent): own selected SWM transfers in lifecycle
