+++ packages/agent/src/dkg-agent-lifecycle.ts
@@ -295,7 +295,6 @@ import {
 } from './sync/requester/shared-memory-sync.js';
 import {
   createSelectedSwmMetaFetcher,
-  SelectedSwmMetaTransferCoordinator,
   type SelectedSwmMetaFetcher,
 } from './sync/selected-swm-meta-fetcher.js';
 import { createSharedMemorySnapshotMaterializer } from './sync/requester/swm-snapshot-materializer.js';
@@ -967,30 +966,6 @@ const selectedSwmMetaRetentionBudgets = new WeakMap<
   object,
   { signature: string; budget: SelectedSwmMetaRetentionBudget }
 >();
-
-const selectedSwmMetaTransferCoordinators = new WeakMap<
-  object,
-  SelectedSwmMetaTransferCoordinator
->();
-
-function selectedSwmMetaTransferCoordinatorFor(
-  owner: object,
-): SelectedSwmMetaTransferCoordinator {
-  let coordinator = selectedSwmMetaTransferCoordinators.get(owner);
-  if (!coordinator) {
-    coordinator = new SelectedSwmMetaTransferCoordinator();
-    selectedSwmMetaTransferCoordinators.set(owner, coordinator);
-  }
-  return coordinator;
-}
-
-/** Release every retained selected-SWM prefix before agent storage teardown. */
-export async function closeSelectedSwmMetaTransfers(owner: object): Promise<void> {
-  const coordinator = selectedSwmMetaTransferCoordinators.get(owner);
-  if (!coordinator) return;
-  selectedSwmMetaTransferCoordinators.delete(owner);
-  await coordinator.close();
-}
 
 function selectedSwmMetaRetentionBudgetFor(
   owner: object,
@@ -6667,7 +6642,7 @@ export class LifecycleSyncMethods extends DKGAgentBase {
 
     return runSyncSingleFlight(this, singleFlightKey, () => (
       selectedSwmEnabled
-        ? selectedSwmMetaTransferCoordinatorFor(this).run(
+        ? this.getSelectedSwmMetaTransfers().run(
           remotePeerId,
           createSelectedMetaFetcher,
           runSync,
