+++ packages/agent/src/dkg-agent.ts
@@ -425,10 +425,7 @@ import { SwmSubstrateMethods } from './dkg-agent-swm-substrate.js';
 import { QueryMethods } from './dkg-agent-query.js';
 import { AgentRegistryMethods } from './dkg-agent-registry.js';
 import { WorkspaceCryptoMethods } from './dkg-agent-crypto.js';
-import {
-  closeSelectedSwmMetaTransfers,
-  LifecycleSyncMethods,
-} from './dkg-agent-lifecycle.js';
+import { LifecycleSyncMethods } from './dkg-agent-lifecycle.js';
 import {
   PublishMethods,
   SEAL_CAPABILITY_GAP_CODE,
@@ -1923,7 +1920,7 @@ export class DKGAgent extends DKGAgentBase {
       // Node stop aborts active transport first; now drain the peer-serial
       // owners and release every retained selected-SWM prefix/checkpoint before
       // the backing store closes.
-      await closeSelectedSwmMetaTransfers(this);
+      await this.closeSelectedSwmMetaTransfers();
     }
     if (this.syncVerifyWorker) {
       await this.syncVerifyWorker.close();
