+++ packages/agent/test/selected-swm-continuation.test.ts
@@ -2,7 +2,6 @@ import { describe, expect, it, vi } from 'vitest';
 import {
   contextGraphWorkspaceMetaGraphUri,
   PROTOCOL_SYNC,
-  type OperationContext,
 } from '@origintrail-official/dkg-core';
 import { workspacePublicQuadsDigest } from '@origintrail-official/dkg-publisher';
 import { OxigraphStore, type Quad } from '@origintrail-official/dkg-storage';
@@ -10,15 +9,11 @@ import type {
   SharedMemorySyncResult,
   SwmSnapshotCoverage,
 } from '../src/dkg-agent-types.js';
-import {
-  closeSelectedSwmMetaTransfers,
-  LifecycleSyncMethods,
-} from '../src/dkg-agent-lifecycle.js';
+import { LifecycleSyncMethods } from '../src/dkg-agent-lifecycle.js';
 import {
   runSelectedSwmContinuations,
 } from '../src/sync/selected-swm-continuation.js';
 import {
-  createSelectedSwmMetaFetcher,
   SelectedSwmMetaTransferCoordinator,
   type SelectedSwmMetaContinuation,
 } from '../src/sync/selected-swm-meta-fetcher.js';
@@ -34,7 +29,6 @@ import {
 } from '../src/sync/requester/page-fetch.js';
 import { estimateQuadHeapBytes } from '../src/sync/memory-telemetry.js';
 import { DURABLE_DATA_SYNC_SESSION_TTL_MS } from '../src/sync/durable-session.js';
-import { createSelectedSwmMetaRetentionBudget } from '../src/sync/selected-swm-meta-budget.js';
 
 const PEER = '12D3KooWSelectedCompleteSwmProvider';
 
@@ -73,20 +67,6 @@ function snapshotManifest(contextGraphId: string, count: number): {
     payloadByRef.set(digest, payload);
   }
   return { meta, payloadByRef };
-}
-
-const selectedMetaTestContext = {
-  operationId: 'selected-meta-owner-test',
-  operationName: 'sync',
-} as OperationContext;
-
-function selectedMetaRetentionBudget() {
-  return createSelectedSwmMetaRetentionBudget({
-    maxRows: 100,
-    maxBytesEstimate: 1024 * 1024,
-    maxPrefixRows: 100,
-    maxPrefixBytesEstimate: 1024 * 1024,
-  });
 }
 
 function cleanDurableResult(): SharedMemorySyncResult {
@@ -241,6 +221,8 @@ interface SelectedProviderSelectionAgent {
     options?: { selectedSwmPriority?: boolean },
   ) => Promise<SharedMemorySyncResult>;
   log: { info: () => void; warn: () => void; debug: () => void };
+  getSelectedSwmMetaTransfers: () => SelectedSwmMetaTransferCoordinator;
+  closeSelectedSwmMetaTransfers: () => Promise<void>;
 }
 
 const callTrySyncFromPeer = LifecycleSyncMethods.prototype.trySyncFromPeer as unknown as (
@@ -413,6 +395,7 @@ function createSelectedSwmLifecycleHarness(
   const metaSinceBatchIds: Array<string | undefined> = [];
   const processedMetaBatches: Quad[][] = [];
   const dateNow = vi.spyOn(Date, 'now').mockImplementation(options.clock.now);
+  let selectedSwmMetaTransfers: SelectedSwmMetaTransferCoordinator | undefined;
 
   const agent: SelectedSwmLifecycleAgentFixture = {
     config: {
@@ -583,6 +566,16 @@ function createSelectedSwmLifecycleHarness(
     contextGraphMetaProjection: { markDirtyFromQuads: () => {} },
     oversizeTombstoneLog: { record: () => {} },
     log: { info: () => {}, warn: () => {}, debug: () => {} },
+    getSelectedSwmMetaTransfers: () => {
+      selectedSwmMetaTransfers ??= new SelectedSwmMetaTransferCoordinator();
+      return selectedSwmMetaTransfers;
+    },
+    closeSelectedSwmMetaTransfers: async () => {
+      const transfers = selectedSwmMetaTransfers;
+      if (!transfers) return;
+      await transfers.close();
+      if (selectedSwmMetaTransfers === transfers) selectedSwmMetaTransfers = undefined;
+    },
   };
 
   return {
@@ -601,162 +594,11 @@ function createSelectedSwmLifecycleHarness(
     },
     close: async () => {
       dateNow.mockRestore();
-      await closeSelectedSwmMetaTransfers(agent);
+      await agent.closeSelectedSwmMetaTransfers();
       await store.close();
     },
   };
 }
-
-describe('selected SWM metadata transfer ownership', () => {
-  it('eagerly releases an expired prefix without requiring another reconciler invocation', async () => {
-    const peerId = 'peer-expiry';
-    const contextGraphId = 'cg-expiry';
-    const baseNow = Date.now();
-    let elapsedMs = 0;
-    const clock = () => baseNow + elapsedMs;
-    const deleteCheckpoint = vi.fn();
-    let ownedFetcher: ReturnType<typeof createSelectedSwmMetaFetcher> | undefined;
-    const coordinator = new SelectedSwmMetaTransferCoordinator({ now: clock });
-
-    const createFetcher = () => {
-      ownedFetcher = createSelectedSwmMetaFetcher({
-        remotePeerId: peerId,
-        requesterScope: 'selected-swm-meta:retained:eager-expiry',
-        retentionBudget: selectedMetaRetentionBudget(),
-        deleteCheckpoint,
-        now: clock,
-        retentionTtlMs: 20,
-        fetchPage: async () => ({
-          quads: [{ subject: 'urn:expiry', predicate: 'urn:p', object: '"o"', graph: 'urn:meta' }],
-          bytesReceived: 1,
-          resumedFromOffset: 0,
-          nextOffset: 1,
-          checkpointKey: 'expiry-checkpoint',
-          completed: false,
-          timedOut: true,
-        }),
-      });
-      return ownedFetcher;
-    };
-
-    try {
-      await coordinator.run(peerId, createFetcher, (fetcher) => fetcher.strategy.fetch({
-        ctx: selectedMetaTestContext,
-        remotePeerId: peerId,
-        contextGraphId,
-        graphUri: 'urn:meta',
-        deadline: clock() + 1_000,
-      }));
-      expect(ownedFetcher?.continuation(contextGraphId).progress).toBe(1);
-
-      elapsedMs = 21;
-      await vi.waitFor(
-        () => expect(ownedFetcher?.continuation(contextGraphId).progress).toBeUndefined(),
-        { timeout: 250 },
-      );
-      expect(deleteCheckpoint).toHaveBeenCalledWith('expiry-checkpoint');
-    } finally {
-      await coordinator.close();
-    }
-  });
-
-  it('expires Context Graph prefixes independently inside one peer owner', async () => {
-    const peerId = 'peer-multi-cg';
-    const baseNow = Date.now();
-    let elapsedMs = 0;
-    const clock = () => baseNow + elapsedMs;
-    let ownedFetcher: ReturnType<typeof createSelectedSwmMetaFetcher> | undefined;
-    const coordinator = new SelectedSwmMetaTransferCoordinator({ now: clock });
-    const createFetcher = () => {
-      ownedFetcher = createSelectedSwmMetaFetcher({
-        remotePeerId: peerId,
-        requesterScope: 'selected-swm-meta:retained:multi-cg',
-        retentionBudget: selectedMetaRetentionBudget(),
-        deleteCheckpoint: () => {},
-        now: clock,
-        retentionTtlMs: 30,
-        fetchPage: async ({ contextGraphId }) => ({
-          quads: [{
-            subject: `urn:${contextGraphId}`,
-            predicate: 'urn:p',
-            object: '"o"',
-            graph: 'urn:meta',
-          }],
-          bytesReceived: 1,
-          resumedFromOffset: 0,
-          nextOffset: 1,
-          checkpointKey: `${contextGraphId}:checkpoint`,
-          completed: false,
-          timedOut: true,
-        }),
-      });
-      return ownedFetcher;
-    };
-    const fetch = (contextGraphId: string) => coordinator.run(
-      peerId,
-      createFetcher,
-      (fetcher) => fetcher.strategy.fetch({
-        ctx: selectedMetaTestContext,
-        remotePeerId: peerId,
-        contextGraphId,
-        graphUri: 'urn:meta',
-        deadline: clock() + 1_000,
-      }),
-    );
-
-    try {
-      const first = await fetch('cg-a');
-      expect(first.result.quads.map((quad) => quad.subject)).toEqual(['urn:cg-a']);
-      elapsedMs = 20;
-      const second = await fetch('cg-b');
-      expect(second.result.quads.map((quad) => quad.subject)).toEqual(['urn:cg-b']);
-
-      elapsedMs = 31;
-      await vi.waitFor(
-        () => expect(ownedFetcher?.continuation('cg-a').progress).toBeUndefined(),
-        { timeout: 250 },
-      );
-      expect(ownedFetcher?.continuation('cg-b').progress).toBe(1);
-    } finally {
-      await coordinator.close();
-    }
-  });
-
-  it('does not serialize independent peers behind one transfer owner', async () => {
-    const coordinator = new SelectedSwmMetaTransferCoordinator();
-    let releasePeerA!: () => void;
-    const peerARelease = new Promise<void>((resolve) => { releasePeerA = resolve; });
-    let signalPeerAStarted!: () => void;
-    const peerAStarted = new Promise<void>((resolve) => { signalPeerAStarted = resolve; });
-    const createEmptyFetcher = (peerId: string) => createSelectedSwmMetaFetcher({
-      remotePeerId: peerId,
-      requesterScope: `selected-swm-meta:retained:${peerId}`,
-      retentionBudget: selectedMetaRetentionBudget(),
-      deleteCheckpoint: () => {},
-      fetchPage: async () => {
-        throw new Error('unused');
-      },
-    });
-
-    try {
-      const peerA = coordinator.run('peer-a', () => createEmptyFetcher('peer-a'), async () => {
-        signalPeerAStarted();
-        await peerARelease;
-      });
-      await peerAStarted;
-      let peerBStarted = false;
-      await coordinator.run('peer-b', () => createEmptyFetcher('peer-b'), async () => {
-        peerBStarted = true;
-      });
-      expect(peerBStarted).toBe(true);
-      releasePeerA();
-      await peerA;
-    } finally {
-      releasePeerA();
-      await coordinator.close();
-    }
-  });
-});
 
 describe('shared-memory freshness classification', () => {
   it('owns producer recovery, bounding, and final classification as one invariant', () => {
@@ -1727,7 +1569,7 @@ describe('selected RFC-64 SWM lifecycle wiring', () => {
         selectedSwmPriority: true,
         sharedMemorySyncPlan: plan,
       });
-      await closeSelectedSwmMetaTransfers(harness.agent);
+      await harness.agent.closeSelectedSwmMetaTransfers();
       await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
         selectedSwmPriority: true,
         sharedMemorySyncPlan: plan,
