+++ packages/agent/vitest.unit.config.ts
@@ -92,7 +92,9 @@ export default defineConfig({
       // #2050 — the stop rule for the bounded repeat of the public SWM peer walk.
       "test/catchup-pass-policy.test.ts",
       "test/selected-swm-continuation.test.ts",
+      "test/selected-swm-meta-transfer-coordinator.test.ts",
       "test/selected-swm-meta-budget.test.ts",
+      "test/outbox-shutdown-lifecycle.test.ts",
       "test/sync-checkpoint-key.test.ts",
       "test/map-with-concurrency.test.ts",
       "test/peer-selection.test.ts",
